Refactored OperationPool to be an interface
This will allow easier changes to operation pools, we're likely to want to provide a bls specific operation pool in a subsequent PR.
